PLANNER-1835 Move classes related to EnvironmentModeTest
