+++ optaplanner-core/src/test/java/org/optaplanner/core/config/solver/testutil/corruptedmoves/TestdataCorruptedUndoMove.java
@@ -1,4 +1,4 @@
-package org.optaplanner.core.impl.testdata.heuristic.move.corrupted.undo;
+package org.optaplanner.core.config.solver.testutil.corruptedmoves;
 
 import org.optaplanner.core.impl.heuristic.move.AbstractMove;
 import org.optaplanner.core.impl.score.director.ScoreDirector;
@@ -6,15 +6,15 @@ import org.optaplanner.core.impl.testdata.domain.TestdataEntity;
 import org.optaplanner.core.impl.testdata.domain.TestdataSolution;
 import org.optaplanner.core.impl.testdata.domain.TestdataValue;
 
-public class TestdataCorruptedEntityUndoMove extends AbstractTestdataMove {
+public class TestdataCorruptedUndoMove extends AbstractTestdataMove {
 
-    public TestdataCorruptedEntityUndoMove(TestdataEntity entity, TestdataValue toValue) {
+    public TestdataCorruptedUndoMove(TestdataEntity entity, TestdataValue toValue) {
         super(entity, toValue);
     }
 
     @Override
     protected AbstractMove<TestdataSolution> createUndoMove(ScoreDirector<TestdataSolution> scoreDirector) {
-        // Corrupts the undo move by creating a new entity and not undo-ing the value
-        return new TestdataCorruptedEntityUndoMove(new TestdataEntity("corrupted"), toValue);
+        // Corrupts the undo move by not undo-ing the value
+        return new TestdataCorruptedEntityUndoMove(entity, toValue);
     }
 }
