+++ optaplanner-core/src/test/java/org/optaplanner/core/config/solver/testutil/corruptedmoves/factory/AbstractTestdataCorruptedUndoMoveTotalMappingFactory.java
@@ -1,4 +1,4 @@
-package org.optaplanner.core.impl.testdata.heuristic.move.corrupted.undo.factory;
+package org.optaplanner.core.config.solver.testutil.corruptedmoves.factory;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -8,9 +8,9 @@ import org.optaplanner.core.impl.heuristic.selector.move.factory.MoveListFactory
 import org.optaplanner.core.impl.testdata.domain.TestdataEntity;
 import org.optaplanner.core.impl.testdata.domain.TestdataSolution;
 import org.optaplanner.core.impl.testdata.domain.TestdataValue;
-import org.optaplanner.core.impl.testdata.heuristic.move.corrupted.undo.AbstractTestdataMove;
-import org.optaplanner.core.impl.testdata.heuristic.move.corrupted.undo.TestdataCorruptedEntityUndoMove;
-import org.optaplanner.core.impl.testdata.heuristic.move.corrupted.undo.TestdataCorruptedUndoMove;
+import org.optaplanner.core.config.solver.testutil.corruptedmoves.AbstractTestdataMove;
+import org.optaplanner.core.config.solver.testutil.corruptedmoves.TestdataCorruptedEntityUndoMove;
+import org.optaplanner.core.config.solver.testutil.corruptedmoves.TestdataCorruptedUndoMove;
 
 public class AbstractTestdataCorruptedUndoMoveTotalMappingFactory implements MoveListFactory<TestdataSolution> {
